+++ CHANGELOG.md
@@ -131,3 +131,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Ameriflux sitename updated. [#302](https://github.com/ncsa/ameriflux-pipeline/issues/302)
 - Deleted unneccessary test modules. [#300](https://github.com/ncsa/ameriflux-pipeline/issues/300)
 - Metmerger modules renamed. [#312](https://github.com/ncsa/ameriflux-pipeline/issues/312)
+- Added key file process in metprocessor. [#316](https://github.com/ncsa/ameriflux-pipeline/issues/316)